Updated model package. Will be refined further in an upcoming refactor.
