+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/controller/EligibilityController.java
@@ -20,10 +20,10 @@ import ca.bc.gov.hlth.hnweb.converter.hl7v2.R15Converter;
 import ca.bc.gov.hlth.hnweb.converter.rapid.R42Converter;
 import ca.bc.gov.hlth.hnweb.converter.rapid.RPBSPPE0Converter;
 import ca.bc.gov.hlth.hnweb.exception.HNWebException;
-import ca.bc.gov.hlth.hnweb.model.CheckEligibilityRequest;
-import ca.bc.gov.hlth.hnweb.model.CheckEligibilityResponse;
-import ca.bc.gov.hlth.hnweb.model.CheckMspCoverageStatusRequest;
-import ca.bc.gov.hlth.hnweb.model.CheckMspCoverageStatusResponse;
+import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityRequest;
+import ca.bc.gov.hlth.hnweb.model.eligibility.CheckEligibilityResponse;
+import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusRequest;
+import ca.bc.gov.hlth.hnweb.model.eligibility.CheckMspCoverageStatusResponse;
 import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnRequest;
 import ca.bc.gov.hlth.hnweb.model.eligibility.InquirePhnResponse;
 import ca.bc.gov.hlth.hnweb.model.eligibility.LookupPhnRequest;
