+++ backend/src/main/java/ca/bc/gov/hlth/hnweb/model/eligibility/CheckEligibilityRequest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.hnweb.model;
+package ca.bc.gov.hlth.hnweb.model.eligibility;
 
 import java.time.LocalDate;
